requirements.txt
google-api-python-client==2.107.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
openai==1.3.5
